Move Navigation to Feature, add Logout
