+++ src/features/navigation/navigation.types.d.ts
@@ -1,0 +1,14 @@
+import { Component } from "vue";
+
+interface IMenuItem {
+	label: string;
+	routerLink?: string;
+	functionCall?: Function;
+	children?: IMenuItem[];
+	icon?: Component;
+}
+
+interface IMenuSection {
+	label: string;
+	children: IMenuItem[];
+}
